+++ Source/Orts.Simulation/Simulator.cs
@@ -1337,7 +1337,7 @@ namespace Orts.Simulation
             else
                 train.TrainMaxSpeedMpS = Math.Min((float)Route.SpeedLimit, conFile.Train.MaxVelocity.A);
 
-            float prevEQres = train.BrakeSystem.EqualReservoirPressurePSIorInHg;
+            double prevEQres = train.BrakeSystem.EqualReservoirPressurePSIorInHg;
             train.AITrainBrakePercent = 100; //<CSComment> This seems a tricky way for the brake modules to test if it is an AI train or not
             train.BrakeSystem.EqualReservoirPressurePSIorInHg = prevEQres; // The previous command modifies EQ reservoir pressure, causing issues with EP brake systems, so restore to prev value
